split UdfProducer class to avoid NoClass exception
